Started level 1
I added cube models of simple objects:

brown: tables
red, green: boxes full of infinite material
grey: turret:
yellow: table to put the turret
black: player

Players can move around with "wasd" keys and pick up items with space. Players can also set items down on tables or put them in the turret (when mounted)

players can also interact with the turret using the "e" key and shoot with space. the "a" and "d" keys are to turn the turret instead of moving when on the mounted turret.

note: there is not crafting system so all items give the turrets the same ammo

note: no animation or models are added

note: added free basic tree models in the assets folder